backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==14.1
python-multipart==0.0.12

# Database
sqlmodel==0.0.22
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Pydantic AI
pydantic-ai==0.0.14
pydantic-ai-mcp==0.0.2

# MCP
mcp==1.1.2

# Utilities
python-dotenv==1.0.1
